+++ changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project are documented in this file.
 
 ## Unreleased
 
+### Bug Fixes
+
+- Fixed PHP Fatal Error for Block Widgets
+
 ## [1.9.11 - 2021-10-23](https://github.com/manzoorwanijk/wptelegram-login/releases/tag/v1.9.11)
 
 ### Bug Fixes